#7 Merge Instruction types into one
- Create instructions.rs
- Add AddInstruction to instructions.rs
- Add CopyInstruction to instructions.rs
- Add ReferenceInstruction to instructions.rs
- Move tests over
