--- src/instructions/mod.rs
@@ -1,4 +1,0 @@
-pub mod add_instruction;
-pub mod copy_instruction;
-pub mod reference_instruction;
-pub mod instruction_enum;
